+++ NEWS.md
@@ -12,6 +12,7 @@
 * Use `option("resgf.indexNode")` to set a default index node universally. Add specification of index node to `resgf_get_fileset()`
 * Add `resgf_get_node_status()` to find out what's going on (programmatically)
 * `resgf_status()` now checks that filesizes agree to within a given threshold.
+* `resgf_get_filelist()` always searches across replicas and versions now
 
 # RESGF 0.1.0